requirements.txt
fastapi
uvicorn
google-generativeai
pymongo
python-dotenv
python-consul2